+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed 
 - Buy button always showing with on hover option.
-### Changed
 - Display buy button on mobile devices.
+### Changed
 - Storefront label names when option is only for web devices.
 
 ## [2.9.12] - 2019-02-01